Move fetch resource restrictions outside of extract_resource_restrictions
