+++ app/domain/authentication/resource_restrictions/extract_resource_restrictions.rb
@@ -7,6 +7,7 @@ module Authentication
       dependencies: {
         resource_restrictions_class: Authentication::ResourceRestrictions::ResourceRestrictions,
         get_restriction_from_annotation: Authentication::ResourceRestrictions::GetRestrictionFromAnnotation,
+        fetch_resource_annotations: Authentication::ResourceRestrictions::FetchResourceAnnotations,
         role_class: ::Role,
         resource_class: ::Resource,
         logger: Rails.logger
@@ -33,28 +34,10 @@ module Authentication
       private
 
       def fetch_resource_annotations
-        resource_annotations
-      end
-
-      def resource_annotations
-        @resource_annotations ||=
-          resource.annotations.each_with_object({}) do |annotation, result|
-            annotation_values = annotation.values
-            value = annotation_values[:value]
-            next if value.blank?
-
-            result[annotation_values[:name]] = value
-          end
-      end
-
-      def resource
-        # Validate role exists, otherwise getting role annotations return empty hash.
-        role_id = @role_class.roleid_from_username(@account, @role_name)
-        resource = @resource_class[role_id]
-
-        raise Errors::Authentication::Security::RoleNotFound, role_id unless resource
-
-        resource
+        @resource_annotations = @fetch_resource_annotations.new.call(
+          account: @account,
+          role_name: @role_name
+        )
       end
 
       def extract_resource_restrictions_from_annotations
@@ -63,7 +46,7 @@ module Authentication
 
       def resource_restrictions_hash
         @resource_restrictions_hash ||=
-          resource_annotations.each_with_object({}) do |(annotation_name, annotation_value), resource_restrictions_hash|
+          @resource_annotations.each_with_object({}) do |(annotation_name, annotation_value), resource_restrictions_hash|
             add_restriction_to_hash(annotation_name, annotation_value, resource_restrictions_hash)
           end
       end
